% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bola_corpus_text}
\alias{bola_corpus_text}
\title{Corpus data non-Leipzig}
\format{A character vector of 18 elements that do not all correspond to one sentence.}
\source{
from BOLA NEWS.com article entitled "Nostalgia Fergie dan Beckham" posted on 16 February 2010
}
\usage{
bola_corpus_text
}
\description{
This is an example input data whose line does not correspond to one sentence as in the Leipzig Corpora.
}
\keyword{datasets}
