% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_leipzig_sample}
\alias{my_leipzig_sample}
\title{Corpus data}
\format{A character vector of 25,000 elements of sentences.}
\source{
\url{http://wortschatz.uni-leipzig.de/en/download}
}
\usage{
my_leipzig_sample
}
\description{
The corpus is a random subset of 25,000 sentences from one of the Indonesian Leipzig Corpora files, i.e., the \code{"ind_news_2008_300K-sentences.txt"}. This corpus file originally contains 300,000 sentences of Indonesian online newspapers.
}
\keyword{datasets}
