% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_pairs.R
\name{word_pairs}
\alias{word_pairs}
\title{Detect word pairs in text}
\usage{
word_pairs(corpus, word_1 = NULL, word_2 = NULL, min_intervening = 0,
  max_intervening = 3)
}
\arguments{
\item{corpus}{A character vector of sentences.}

\item{word_1}{A regular expressions for the first word. The regex must enclose the word with word boundary character (i.e. \code{"\\\\b"}).}

\item{word_2}{A regular expressions for the second word. The regex must enclose the word with word boundary character (i.e. \code{"\\\\b"}).}

\item{min_intervening}{Number of minimum occurrence of the intervening word.}

\item{max_intervening}{Number of minimum occurrence of the intervening word.}
}
\value{
A list object with the following elements:
\itemize{
\item{\code{pattern}: the extracted pattern spanning from the first word to the second word.}
\item{\code{pattern_tagged}: the version of \code{pattern} containing tags for the first and the second word.}
\item{\code{matches}: the sentence matches containing the word pairs that are tagged for the first and the second word.}
}
}
\description{
\code{word_pairs} searches for the occurrences of a pair of words in sentences. These words can be separated by intervening strings (viz. other in-between words), which is also handled by \code{word_pairs}.
}
\examples{
word_1 <- "\\\\bmenanyakan\\\\b"
word_2 <- "\\\\bkepada\\\\b"
corpus <- my_leipzig_sample
m <- word_pairs(corpus, word_1 = word_1, word_2 = word_2, min_intervening = 0, max_intervening = 3)
}
\references{
Rajeg, Gede Primahadi Wijaya. (2018). wordpairs: An R package to retrieve word pair in sentences of the (Indonesian) Leipzig Corpora.
}
